+++ frontend/src/_services/utils/types/api.ts
@@ -1,0 +1,3 @@
+import { HTTP_METHODS } from "@/services/utils/constants";
+
+export type HttpMethodType = (typeof HTTP_METHODS)[keyof typeof HTTP_METHODS];